requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-MySQLdb==2.0.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mercadopago==2.2.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3